Oh fuck yes.
Let me take a minute to explain right here why I love ember so much.

Does it take a long time to learn? Yes. But that isn't because ember is poorly put together, it is because ember is trying to do three things at once. First it is trying to get input-data-and-display synced together so that there is a consistent application. Second it is trying to standardize json apis and the code that syncronizes local and server state. Third, it is trying to do this with as little code, boilerplate or otherwise.

I'm still 100% convinced that it is worth putting the time into learning ember. While I still find JavaScript a frustrating language, ember is a great framework, and I'm hating client side less and less.